+++ config/settings/test.py
@@ -31,6 +31,7 @@ EMAIL_PORT = 1025
 # In-memory email backend stores messages in django.core.mail.outbox
 # for unit testing purposes
 EMAIL_BACKEND = 'django.core.mail.backends.locmem.EmailBackend'
+EMAIL_SENDER = 'noreply@example.com'
 
 # CACHING
 # ------------------------------------------------------------------------------